backend_server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
japanize-matplotlib==1.1.3
pydantic==2.5.0
mysql-connector-python==8.2.0
requests==2.31.0
